server/requirements.txt
tensorflow==1.15
connexion[swagger-ui]==2.7.0
pandas>=1.1.3
pyneuroner[cpu]==1.0.8
python_dateutil==2.8.1
swagger-ui-bundle==0.0.8